+++ components/ui/bubbleMenu.vue
@@ -66,11 +66,11 @@
           </button>
 
                     <!-- AI Button -->
-                    <button @click="openAIBar" class=" hover:dark:bg-[#171717] hover:bg-gray-100 p-2 px-2" aria-label="AI Edit" title="Write with AI.">
+                    <!-- <button @click="openAIBar" class=" hover:dark:bg-[#171717] hover:bg-gray-100 p-2 px-2" aria-label="AI Edit" title="Write with AI.">
   
                       <svg xmlns="http://www.w3.org/2000/svg" width="23" viewBox="0 0 24 24"><path fill="currentColor" d="M5.838 10.184c.19.49.64.81 1.17.81s.99-.32 1.17-.81l.71-1.92c.02-.07.07-.12.14-.14l1.92-.71c.49-.19.81-.64.81-1.17s-.32-.99-.81-1.17l-1.92-.71a.24.24 0 0 1-.14-.14l-.71-1.92c-.19-.49-.64-.81-1.17-.81s-.99.32-1.17.81l-.71 1.92c-.02.07-.07.12-.14.14l-1.92.71c-.49.19-.81.64-.81 1.17s.32.99.81 1.17l1.92.71c.07.03.12.08.14.14zm-.33-3.47l-1.27-.47l1.27-.47c.48-.18.85-.55 1.03-1.03l.47-1.27l.47 1.27c.18.47.56.85 1.03 1.02l1.27.47l-1.27.47c-.48.18-.85.55-1.03 1.03l-.47 1.27l-.47-1.27c-.18-.47-.56-.85-1.03-1.02m-.25 15.53c0 .41.34.75.75.75h1.5l-.01.01c1.27 0 1.97 0 2.65-.28s1.17-.77 2.07-1.67l8.15-8.15l.046-.046c.541-.55.85-.864 1.044-1.253c.38-.76.38-1.66 0-2.42c-.182-.393-.483-.694-1.018-1.228l-.072-.072l-.047-.046c-.55-.542-.864-.85-1.253-1.044c-.76-.38-1.66-.38-2.42 0c-.393.182-.694.483-1.228 1.017l-.072.073l-1.48 1.48l-6.66 6.66c-.89.89-1.39 1.39-1.67 2.07s-.28 1.38-.28 2.65zm1.51-.75v-.74c0-1.12 0-1.67.17-2.08s.56-.8 1.35-1.59l6.14-6.14l.95.95c.29.29.77.29 1.06 0s.29-.77 0-1.06l-.95-.95l.95-.95c.45-.45.72-.72.9-.81c.34-.17.75-.17 1.09 0c.18.09.45.36.9.81s.72.72.81.9c.17.34.17.75 0 1.09c-.09.18-.36.45-.81.9l-8.15 8.15c-.84.84-1.2 1.19-1.59 1.35c-.41.17-.97.17-2.08.17z" color="currentColor"/></svg>
 
-</button>
+</button> -->
 
           <button @click="readSelectedText"
             :class="{ 'is-reading': isReading, 'hover:dark:bg-[#171717] hover:bg-gray-100': true, 'p-2 px-2 border-l border-gray-200 dark:border-[#525252]': true }"
