+++ src-tauri/tauri.conf.json
@@ -7,7 +7,7 @@
   },
   "package": {
     "productName": "fylepad",
-    "version": "2.2.0"
+    "version": "2.3.0"
   },
   "tauri": {
     "allowlist": {
